Add the load API
Should be executed when opening a project.
